Debugged c89 related error

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,14 +11,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	unsigned long i;
 
 	if (array == NULL)
 		return (-1);
 
 	for (i = 0; i < size; ++i)
 	{
-		printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
